``` refactor: simplify test imports and improve code formatting in e2e tests
- Updated imports in e2e tests for `dataIntegrity.test.ts` and `userJourney.test.ts` to use a consistent relative path structure.
- Standardized code formatting for better readability (e.g., fixed indentation, eliminated trailing spaces, improved readability in chained methods).
- No functional changes made; purely organizational and readability refactoring.
```
